Fix added whitespace to config values
Atom was adding whitespace to my config values at the end of the string which was breaking the build command for the linter. So I just added a trim to the output which fixes it.

Note that I didn't have a white space as an extra character in the config value, Atom was adding it.